requirements.txt
python-dotenv 
anthropic 
tiktoken
